Added combat code, modified the Death scene, the LaserWeapon scene and the EscapePod scene.
Combat code will be used to change scenes CentralCorridor and TheBridge.
Death scene will now give the option to restart at the last scene.
LaserWeapon scene now gives feedback on the code choices to help with the guessing.
EscapePod scene now changed so it is easier to guess the right pod.

	modified:   ex43 notes
	modified:   ex43_book.py
	modified:   ex43_book_test.py